+++ CKS3Connection.h
@@ -40,7 +40,7 @@
 	unsigned long long	bytesTransferred;
 	unsigned long long	bytesToTransfer;
 	unsigned long long	transferHeaderLength;
-	unsigned int myLastPercent;
+	NSUInteger myLastPercent;
 	NSFileHandle *myDownloadHandle;
     
 @private
